MAP-1414 Replace maxCapacity with actualCapacity in cell selection (#83)
* MAP-1414 Replace maxCapacity with actualCapacity in cell selection

This change updates the cell selection process to prefer 'actualCapacity', which considers both 'workingCapacity' and 'maxCapacity', over 'maxCapacity' alone. This makes sure that any operational changes to cell capacity are reflected in selection process. The updates were made in the relevant controller, view, and test files.

* MAP-1414 Filter out prisoner's current cell in cell selection

The code difference introduces a filter in the selection of cells to exclude the current cell of the prisoner. Previously, the selectedCells array included the prisoner's own cell, which is not necessary as the purpose is to find potential cells to move the prisoner to, excluding their current one. This not only delivers a more accurate list of cells but also reduces the potential for user selection errors.

* MAP-1414 Filter out prisoner's current cell in cell selection